refactor(infra): extract the config section appliers
Four apply*Section methods — roughly 200 lines of database/redis/storage/engine mapping — sat inside
the controller although each already took everything it needed as parameters. They move to
config-sections.ts as free functions. applyEngineSection is the only one that read instance state
(the engine registry, to validate a submitted engine id) and now takes the factory as its fourth
argument, still called lazily behind the same guard.

Pure code motion: the 57 existing controller cases pass unedited. The context object is still
mutated in place rather than replaced, so the injection guard and the merge continue to inspect the
same object.
